增加删除和更新论文的功能

+++ src/main/java/com/database/paperms/controller/PaperController.java
@@ -7,6 +7,7 @@ import com.database.paperms.response.ResultData;
 import com.database.paperms.response.ReturnCode;
 import com.database.paperms.service.PaperService;
 import com.database.paperms.utils.RegexUtil;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -21,8 +22,14 @@ public class PaperController {
     @PostMapping("/add")
     public ResultData add(@RequestBody Paper paper){
         try {
-            paperService.savePaper(paper);
-            return ResultData.success();
+            if(paper.getPaperId() != null) {
+                if (paperService.getPaper(paper.getPaperId()) == null) {
+                    paperService.savePaper(paper);
+                    return ResultData.success();
+                } else {
+                    return ResultData.fail(ReturnCode.USED_PAPER_ID);
+                }
+            }
         }catch(ClassCastException e){
             e.printStackTrace();
         }
@@ -31,24 +38,22 @@ public class PaperController {
 
     @DeleteMapping("/delete/{id}")
     public ResultData delete(@PathVariable("id") Integer paperId){
-        try {
+        if(paperService.getPaper(paperId) != null){
             paperService.deletePaper(paperId);
             return ResultData.success();
-        }catch(ClassCastException e){
-            e.printStackTrace();
+        }else{
+            return ResultData.fail(ReturnCode.NOT_EXISTENT_PAPER_ID);
         }
-        return ResultData.fail(ReturnCode.CLASS_CAST_ERROR);
     }
 
     @PostMapping("/update")
     public ResultData update(@RequestBody Paper paper){
-        try {
+        if(paperService.getPaper(paper.getPaperId()) != null){
             paperService.updatePaper(paper);
             return ResultData.success();
-        }catch(ClassCastException e){
-            e.printStackTrace();
+        }else{
+            return ResultData.fail(ReturnCode.NOT_EXISTENT_PAPER_ID);
         }
-        return ResultData.fail(ReturnCode.CLASS_CAST_ERROR);
     }
